requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0
python-dotenv==1.0.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
transformers==4.35.2
torch==2.5.1
pdfplumber==0.10.3
requests==2.31.0
httpx==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
